+++ services/gcloud_push/Cargo.toml
@@ -10,6 +10,7 @@ google-pubsub1_beta2 = "1.0.8+20181001"
 envy = "0.3"
 redis = "0.9"
 redis_context = { path = "../redis_context" }
+redis_delta = { path = "../redis_delta" }
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
